Refactor renderers into own directory
